requirements.txt
Flask==3.0.0
wechatpy==1.8.18
cryptography>=3.0.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.2.1
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
factory-boy==3.3.0
pydantic==2.12.5
pydantic-settings==2.12.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Migrate==4.1.0
sentry-sdk[flask]==2.49.0
